backend/requirements.txt
Flask==3.0.3
Werkzeug==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
Pillow==10.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
langdetect==1.0.9
requests==2.32.3
reportlab==4.1.0
gTTS==2.5.1
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==1.0.0
Flask-Migrate==4.0.5
pytest==8.2.2
pytest-mock==3.14.0